+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## Version 100107.01
+
+- Added 10.1.7 support.
+
+### Blizzard
+
+- Fixed an issue where populated gem sockets were sometimes shown as empty.
+
+### Minimap
+
+- Added 125% and 150% minimap size options.
+
 ## Version 100105.01
 
 - Added 10.1.5 support.